change adapter to linear layout

+++ app/src/main/java/jason/github/com/photofans/utils/DisplayUtil.java
@@ -28,7 +28,7 @@ public class DisplayUtil {
     public static double getScreenRatio(){
         DisplayMetrics metrics = getScreenDimen();
 
-        return metrics.widthPixels/metrics.heightPixels;
+        return ((double)metrics.widthPixels)/metrics.heightPixels;
     }
 
     public static int dp2px(int px){
